ported all c functionality (now with arg parsing!)
inb4 "rewrite it in rust" :P

but really, I thought it would be nice to ease myself back into
rust development before doing it for the next 4 months at Google.
plus, I _really_ didn't want to write commandline parsing in pure
C, and clap makes argument parsing really nice.

with this basic functionality implmeneted, i'll probably take a
stab at reverse-engineering some of the easier protocols (i.e: how
to update the key-brightness without resending the full config)

Long term goal: take a crack at dumping the keyboard firmware, and
writing some custom presets :O. It's a tall order (especially for
a newb like me), but I've seen others do it with other keyboards,
so it _should_ be possible. the keyboard uses a ARM microcontroller
for controlling the lighting effects, so if I can dump the firmware
and tweak it a bit, i _should_ be able to do some neat stuff!
(snake directly on the keyboard O_o???)

p.s: i'm in the middle of exam season. probably not the best time
to start on a project like this.
whoops.